webapp/requirements.txt
fastapi
uvicorn
jinja2
pandas
numpy
matplotlib
seaborn
scikit-learn
fpdf2
openpyxl
python-multipart
